+++ src/accessory/BaseAccessory.ts
@@ -110,7 +110,9 @@ class BaseAccessory {
   }
 
   configureStatusActive() {
-    if ((this.device.parent_id || this.device.infrared_ac_command_mode === 'device-sharing')
+    if ((this.device.parent_id
+      || this.device.infrared_ac_command_mode === 'device-sharing'
+      || this.device.infrared_remote_command_mode === 'device-sharing')
       && this.device.isIRRemoteControl()) {
       for (const service of this.accessory.services) {
         if (service.testCharacteristic(this.Characteristic.StatusActive)) {